Split module graph runtime implementation
